Restructured tests to have a test file for each module Restructed modules to have a module file for each table
